key-free/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
jinja2==3.1.4
python-dotenv==1.0.1
itsdangerous==2.2.0
sqlmodel==0.0.19
aiosqlite==0.20.0
python-multipart==0.0.9
